+++ src/AzzyBot.Bot/Services/Modules/AzuraCastFileService.cs
@@ -10,8 +10,8 @@ using AzzyBot.Bot.Utilities.Records.AzuraCast;
 using AzzyBot.Core.Logging;
 using AzzyBot.Core.Utilities;
 using AzzyBot.Core.Utilities.Encryption;
-using AzzyBot.Data;
 using AzzyBot.Data.Entities;
+using AzzyBot.Data.Services;
 using DSharpPlus.Entities;
 using Microsoft.Extensions.Logging;