+++ src/AzzyBot.Bot/Services/Modules/AzuraCastUpdateService.cs
@@ -6,8 +6,8 @@ using AzzyBot.Bot.Resources;
 using AzzyBot.Bot.Utilities;
 using AzzyBot.Bot.Utilities.Records.AzuraCast;
 using AzzyBot.Core.Utilities.Encryption;
-using AzzyBot.Data;
 using AzzyBot.Data.Entities;
+using AzzyBot.Data.Services;
 using DSharpPlus.Entities;
 
 namespace AzzyBot.Bot.Services.Modules;