+++ src/AzzyBot.Bot/Settings/Validators/AppStatsValidator.cs
@@ -1,8 +1,10 @@
-﻿using AzzyBot.Core.Utilities.Records;
+﻿using System.Diagnostics.CodeAnalysis;
+using AzzyBot.Core.Utilities.Records;
 using Microsoft.Extensions.Options;
 
 namespace AzzyBot.Bot.Settings.Validators;
 
+[SuppressMessage("Roslynator", "RCS1251:Remove unnecessary braces from record declaration", Justification = "Class has to be empty.")]
 [OptionsValidator]
 public sealed partial class AppStatsValidator : IValidateOptions<AppStats>
 {