+++ src/AzzyBot.Data/Migrations/20241223142128_AddLegals.Designer.cs
@@ -12,15 +12,15 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace AzzyBot.Data.Migrations
 {
     [DbContext(typeof(AzzyDbContext))]
-    [Migration("20240804151746_InitialCreation")]
-    partial class InitialCreation
+    [Migration("20241223142128_AddLegals")]
+    partial class AddLegals
     {
         /// <inheritdoc />
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "8.0.7")
+                .HasAnnotation("ProductVersion", "9.0.0")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
@@ -36,10 +36,10 @@ namespace AzzyBot.Data.Migrations
                     b.Property<int>("AzuraCastId")
                         .HasColumnType("integer");
 
-                    b.Property<DateTime>("LastServerStatusCheck")
+                    b.Property<DateTimeOffset>("LastServerStatusCheck")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<DateTime>("LastUpdateCheck")
+                    b.Property<DateTimeOffset>("LastUpdateCheck")
                         .HasColumnType("timestamp with time zone");
 
                     b.Property<bool>("ServerStatus")
@@ -131,7 +131,7 @@ namespace AzzyBot.Data.Migrations
                     b.Property<bool>("FileChanges")
                         .HasColumnType("boolean");
 
-                    b.Property<DateTime>("LastFileChangesCheck")
+                    b.Property<DateTimeOffset>("LastFileChangesCheck")
                         .HasColumnType("timestamp with time zone");
 
                     b.Property<int>("StationId")
@@ -160,10 +160,10 @@ namespace AzzyBot.Data.Migrations
                     b.Property<int>("AzuraCastId")
                         .HasColumnType("integer");
 
-                    b.Property<DateTime>("LastRequestTime")
+                    b.Property<DateTimeOffset>("LastRequestTime")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<DateTime>("LastSkipTime")
+                    b.Property<DateTimeOffset>("LastSkipTime")
                         .HasColumnType("timestamp with time zone");
 
                     b.Property<int>("StationId")
@@ -214,6 +214,34 @@ namespace AzzyBot.Data.Migrations
                     b.ToTable("AzuraCastStationPreferences");
                 });
 
+            modelBuilder.Entity("AzzyBot.Data.Entities.AzuraCastStationRequestEntity", b =>
+                {
+                    b.Property<int>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("integer");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+
+                    b.Property<bool>("IsInternal")
+                        .HasColumnType("boolean");
+
+                    b.Property<string>("SongId")
+                        .IsRequired()
+                        .HasColumnType("text");
+
+                    b.Property<int>("StationId")
+                        .HasColumnType("integer");
+
+                    b.Property<DateTimeOffset>("Timestamp")
+                        .HasColumnType("timestamp with time zone");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("StationId");
+
+                    b.ToTable("AzuraCastStationRequests");
+                });
+
             modelBuilder.Entity("AzzyBot.Data.Entities.GuildEntity", b =>
                 {
                     b.Property<int>("Id")
@@ -223,6 +251,12 @@ namespace AzzyBot.Data.Migrations
                     NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
 
                     b.Property<bool>("ConfigSet")
+                        .HasColumnType("boolean");
+
+                    b.Property<DateTimeOffset>("LastPermissionCheck")
+                        .HasColumnType("timestamp with time zone");
+
+                    b.Property<bool>("LegalsAccepted")
                         .HasColumnType("boolean");
 
                     b.Property<decimal>("UniqueId")
@@ -327,6 +361,17 @@ namespace AzzyBot.Data.Migrations
                     b.Navigation("Station");
                 });
 
+            modelBuilder.Entity("AzzyBot.Data.Entities.AzuraCastStationRequestEntity", b =>
+                {
+                    b.HasOne("AzzyBot.Data.Entities.AzuraCastStationEntity", "Station")
+                        .WithMany("Requests")
+                        .HasForeignKey("StationId")
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
+
+                    b.Navigation("Station");
+                });
+
             modelBuilder.Entity("AzzyBot.Data.Entities.GuildPreferencesEntity", b =>
                 {
                     b.HasOne("AzzyBot.Data.Entities.GuildEntity", "Guild")
@@ -356,6 +401,8 @@ namespace AzzyBot.Data.Migrations
 
                     b.Navigation("Preferences")
                         .IsRequired();
+
+                    b.Navigation("Requests");
                 });
 
             modelBuilder.Entity("AzzyBot.Data.Entities.GuildEntity", b =>